testing

+++ atomic-core/partial-mngr/file-move.php
@@ -4,36 +4,42 @@
 
 if (!empty($_POST['moveFile'])){
 
-//Set user input variable
 
-$fileName = test_input($_POST["fileName"]);
-$catName = test_input($_POST["compDir"]);
-$newDir = test_input($_POST["newDir"]);
+  $fileName = test_input($_POST["fileName"]);
+  $catName = test_input($_POST["compDir"]);
+  $newDir = test_input($_POST["newDir"]);
 
 
-  moveChangeCommentString($catName, $fileName, $newDir);
+ /* $filename = '../components/'.$newDir.'/'.$fileName.'.php';
 
-  moveScssFile($catName, $fileName, $newDir);
-  
-  deleteScssImportString($catName, $fileName );
-  
-  writeScssImportFile($newDir, $fileName );
-  
-  moveCompFile($catName, $fileName, $newDir);
-  
-  deleteCompIncludetString($catName, $fileName );
-  
-  createIncludeString($newDir, $fileName );
-  
-  //moveAjaxFile($catName, $fileName, $newDir);
-  
-  deleteAjaxCompFile($catName, $fileName);
-  
-  createAjaxIncludeAndCompFile($newDir, $fileName);
-  
-  
-  header("location: http://127.0.0.1/atomic-docs/atomic-core/$newDir.php");
-}
+  if (file_exists($filename)) {
+      echo "The file $filename exists";
+  } else {
+      echo "The file $filename does not exist";*/
+
+
+    moveChangeCommentString($catName, $fileName, $newDir);
+
+    moveScssFile($catName, $fileName, $newDir);
+    
+    deleteScssImportString($catName, $fileName );
+    
+    writeScssImportFile($newDir, $fileName );
+    
+    moveCompFile($catName, $fileName, $newDir);
+    
+    deleteCompIncludetString($catName, $fileName );
+    
+    createIncludeString($newDir, $fileName );
+    
+    deleteAjaxCompFile($catName, $fileName);
+    
+    createAjaxIncludeAndCompFile($newDir, $fileName);
+    
+    
+    header("location: http://127.0.0.1/atomic-docs/atomic-core/$newDir.php");
+
+  }
 
 
 ?>
